peersupport/requirements.txt
transformers==4.43.3
peft==0.11.1
torch==2.3.1
python-dotenv==1.0.1
openai==1.43.0
loguru==0.7.2
sqlalchemy==2.0.32
pytz==2024.1
apscheduler==3.10.4
langchain==0.2.13
langgraph==0.2.34
discord.py==2.4.0
emoji==0.6.0




